# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API Description",
        "contact": {
            "name": "Your Name",
            "email": "you@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "summary": "Get all products",
                "description": "Endpoint to retrieve a list of all products.",
                "operationId": "5c5bbec0ba3ada1a2cbdc47997f9f857",
                "responses": {
                    "200": {
                        "description": "List of products"
                    }
                }
            },
            "post": {
                "tags": [
                    "products"
                ],
                "summary": "Create a new product",
                "description": "Endpoint to create a new products.",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The title of the product",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "The description of the product",
                                        "type": "text"
                                    },
                                    "price": {
                                        "description": "The price of the product",
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "description": "The roles to which the product is targeted",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "updated_at": {
                                        "description": "The roles to which the product is targeted",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification created successfully"
                    },
                    "400": {
                        "description": "Bad request, validation failed"
                    }
                }
            }
        },
        "/api/products/{product_id}": {
            "put": {
                "tags": [
                    "products"
                ],
                "summary": "Update product details",
                "description": "Endpoint to update product details.",
                "operationId": "e3dc107d1eeb91b58436b562af0648b3",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID of the product to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "The title of the product",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "The description of the product",
                                        "type": "text"
                                    },
                                    "price": {
                                        "description": "The price of the product",
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "description": "The roles to which the product is targeted",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "updated_at": {
                                        "description": "The roles to which the product is targeted",
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "product details updated successfully"
                    },
                    "400": {
                        "description": "Bad request, validation failed"
                    }
                }
            },
            "delete": {
                "tags": [
                    "products"
                ],
                "summary": "Delete a product",
                "description": "Endpoint to delete a product by ID.",
                "operationId": "1e2ccb4ad3905276818a481e066ba14e",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID of the product to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product deleted successfully"
                    },
                    "404": {
                        "description": "product not found"
                    }
                }
            }
        },
...
    }
}